Dramatic memory increase; native ops now in Flash.
The code now looks up symbols in two passes: once inside
the global `c_ops` stored in Flash, and once in the `Forth::_dict`.
The result is dramatic; even after reserving 280 bytes
for the CPU stack, we still have 1.4K available for our Forth code.
We are making excellent use of our 2K of SRAM :-)